fix a bug of min max scores

+++ bid_aware_ann_post_filtering/post_gpu.cu
@@ -252,8 +252,8 @@ vector<vector<ReqDocPair>> postGpuAlgoBatch(const vector<CentroidCpu> &centroids
                                             const vector<ItemCpu> &docs,
                                             int numToRetrieve,
                                             bool enableBidAware,
-                                            int minScore,
-                                            int maxScore)
+                                            float minScore,
+                                            float maxScore)
 {
     assert(minScore < maxScore);
