+++ bid_aware_ann_post_filtering/post_gpu.cuh
@@ -21,7 +21,7 @@ vector<vector<ReqDocPair>> postGpuAlgoBatch(const vector<CentroidCpu> &centroids
                                             const vector<ItemCpu> &docs,
                                             int numToRetrieve,
                                             bool enableBidAware,
-                                            int minScore,
-                                            int maxScore);
+                                            float minScore,
+                                            float maxScore);
 
 #endif // POST_GPU_CUH
